Optimized, -3 bytes, -4cc to 9cc (avg -6.5cc)
In the first section, we are essentially selecting between toggling the top bit of `E` or doing `sbc hl,de`. Since we eventually want `E` in `A`, anyways, I set `A` to `0`, rotate in the carry to make `A` either `0x00` or `0x80`. Then I add it to E, essentially loading `E` into `A` and toggling the top bit at the same time if necessary.

In the second section, the `and 0FDh` is just resetting bit 1 of `A` (which is guaranteed to be 1 at this step). In the original, bit 0 is also 1, so we could replace this with a `sub 2` in the case that carry is set. However, if we don't increment the `A` before storing it to `E`, then since we discard the bottom bit either way, we can just subtract 1 if carry is set, 0 otherwise.